fastapi/fastapi_multi_tenant/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pydantic>=2.3.0
pydantic-settings>=2.0.0
alembic>=1.12.0
email-validator>=2.0.0
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.24.0
mysql-connector-python>=9.3.0
